initialize crosschain refactoring
