+++ x/crosschain/keeper/msg_server_update_tss.go
@@ -2,21 +2,25 @@ package keeper
 
 import (
 	"context"
+	authoritytypes "github.com/zeta-chain/zetacore/x/authority/types"
 
 	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
-	observerTypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
 
-// Authorized: admin policy group 2.
+// UpdateTssAddress updates the TSS address.
 func (k msgServer) UpdateTssAddress(goCtx context.Context, msg *types.MsgUpdateTssAddress) (*types.MsgUpdateTssAddressResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	// check if authorized
 	// TODO : Add a new policy type for updating the TSS address
-	if msg.Creator != k.zetaObserverKeeper.GetParams(ctx).GetAdminPolicyAccount(observerTypes.Policy_Type_group2) {
+	// https://github.com/zeta-chain/node/issues/1715
+	if !k.GetAuthorityKeeper().IsAuthorized(ctx, msg.Creator, authoritytypes.PolicyType_groupAdmin) {
 		return nil, errorsmod.Wrap(sdkerrors.ErrUnauthorized, "Update can only be executed by the correct policy account")
 	}
+
 	currentTss, found := k.zetaObserverKeeper.GetTSS(ctx)
 	if !found {
 		return nil, errorsmod.Wrap(types.ErrUnableToUpdateTss, "cannot find current TSS")
@@ -24,6 +28,7 @@ func (k msgServer) UpdateTssAddress(goCtx context.Context, msg *types.MsgUpdateT
 	if currentTss.TssPubkey == msg.TssPubkey {
 		return nil, errorsmod.Wrap(types.ErrUnableToUpdateTss, "no new tss address has been generated")
 	}
+
 	tss, ok := k.zetaObserverKeeper.CheckIfTssPubkeyHasBeenGenerated(ctx, msg.TssPubkey)
 	if !ok {
 		return nil, errorsmod.Wrap(types.ErrUnableToUpdateTss, "tss pubkey has not been generated")
@@ -34,6 +39,7 @@ func (k msgServer) UpdateTssAddress(goCtx context.Context, msg *types.MsgUpdateT
 	if len(k.zetaObserverKeeper.GetSupportedChains(ctx)) != len(tssMigrators) {
 		return nil, errorsmod.Wrap(types.ErrUnableToUpdateTss, "cannot update tss address not enough migrations have been created and completed")
 	}
+
 	// GetAllTssFundMigrators would return the migrators created for the current migration
 	// if any of the migrations is still pending we should not allow the tss address to be updated
 	// we can wait for all migrations to complete before updating; this includes btc and eth chains.
@@ -50,6 +56,7 @@ func (k msgServer) UpdateTssAddress(goCtx context.Context, msg *types.MsgUpdateT
 	}
 
 	k.GetObserverKeeper().SetTssAndUpdateNonce(ctx, tss)
+
 	// Remove all migrators once the tss address has been updated successfully,
 	// A new set of migrators will be created when the next migration is triggered
 	k.zetaObserverKeeper.RemoveAllExistingMigrators(ctx)
